+++ packages/core/types/index.d.ts
@@ -1,14 +1,15 @@
 export { default as env } from './env';
 export type { LoadingState, UNListener } from './utils';
 export { buildConfigSetter, deepClone, deepMerge, isPromise, isServer, SingleDispatcher, toPromise } from './utils';
-export type { Action, ActionCreator, ActionError, AsyncEluxComponent, CommonModel, CommonModelClass, CommonModule, Dispatch, EluxComponent, EluxContext, EluxStoreContext, GetState, IAppRender, IRouter, IRouteRecord, IStore, VStore, Location, ModelAsCreators, ModuleGetter, ModuleState, RouteAction, RouteEvent, RouterInitOptions, RouteRuntime, RouteTarget, StoreLogger, StoreLoggerInfo, StoreMiddleware, StoreState, } from './basic';
-export { coreConfig, getClientRouter, setCoreConfig } from './basic';
+export type { Action, ActionCreator, ActionError, AsyncEluxComponent, CommonModel, CommonModelClass, CommonModule, Dispatch, EluxComponent, EluxContext, EluxStoreContext, GetState, IAppRender, IRouter, IRouteRecord, IStore, Location, ModelAsCreators, ModuleGetter, ModuleState, RouteAction, RouteEvent, RouterInitOptions, RouteRuntime, RouteTarget, StoreLogger, StoreLoggerInfo, StoreMiddleware, StoreState, VStore, } from './basic';
+export { coreConfig, ErrorCodes, getClientRouter, setCoreConfig, isMutable } from './basic';
 export { errorAction, setProcessedError } from './actions';
 export { getComponent, getEntryComponent, getModule, getModuleApiMap, injectComponent, injectModule, moduleExists } from './inject';
 export { effect, effectLogger, EmptyModel, exportComponent, exportView, reducer, setLoading } from './module';
-export { CoreRouter, modelHotReplacement, Store } from './store';
+export { modelHotReplacement, Store } from './store';
 export type { API, Facade, GetPromiseComponent, GetPromiseModule, HandlerToAction, IGetComponent, IGetData, ILoadComponent, ModuleFacade, PickModelActions, PickThisActions, ReturnComponents, } from './facade';
 export { BaseModel, exportModule, getApi } from './facade';
 export type { RenderOptions } from './app';
 export { buildApp, buildProvider, buildSSR, getTplInSSR } from './app';
+export { BaseNativeRouter, locationToNativeLocation, locationToUrl, nativeLocationToLocation, nativeUrlToUrl, Router, urlToLocation, urlToNativeUrl, } from './route';
 //# sourceMappingURL=index.d.ts.map