+++ packages/model/rollup.config.js
@@ -1,0 +1,3 @@
+import build from '../../rollup.build';
+const config = build(__dirname);
+export default config;