+++ packages/model/types/index.d.ts
@@ -1,0 +1,2 @@
+export * from '@elux/app';
+//# sourceMappingURL=index.d.ts.map