+++ packages/react-components/src/Link.tsx
@@ -1,5 +1,4 @@
-import {coreConfig, IRouteRecord, Location, RouteAction, RouteTarget} from '@elux/core';
-import {urlToNativeUrl} from '@elux/route';
+import {coreConfig, IRouteRecord, Location, RouteAction, RouteTarget, urlToNativeUrl} from '@elux/core';
 import {FC, HTMLAttributes, MouseEvent, useCallback, useMemo, useRef} from 'react';
 
 /**