+++ packages/react-components/src/index.ts
@@ -1,4 +1,5 @@
 import {setCoreConfig} from '@elux/core';
+
 import AppRender from './App';
 import {UseRouter} from './base';
 import {LoadComponent, LoadComponentOnError, LoadComponentOnLoading} from './LoadComponent';