+++ packages/vue-components/src/index.ts
@@ -1,5 +1,7 @@
-import {setCoreConfig} from '@elux/core';
 import {reactive} from 'vue';
+
+import {setCoreConfig} from '@elux/core';
+
 import AppRender from './App';
 import {UseRouter, UseStore} from './base';
 import {LoadComponent, LoadComponentOnError, LoadComponentOnLoading} from './LoadComponent';
@@ -15,7 +17,7 @@ setCoreConfig({
   LoadComponentOnLoading,
 });
 
-export {setVueComponentsConfig, connectStore} from './base';
+export {connectStore, setVueComponentsConfig} from './base';
 export {EWindow} from './EWindow';
 export {RouterComponent} from './Router';
 export {DocumentHead} from './DocumentHead';