+++ packages/taro/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/hiisea/elux/compare/v2.6.0-alpha.4...v2.6.0) (2022-08-02)
+
+**Note:** Version bump only for package @elux/taro
+
+
+
+
+
+# [2.6.0-alpha.2](https://github.com/hiisea/elux/compare/v2.6.0-alpha.1...v2.6.0-alpha.2) (2022-08-02)
+
+**Note:** Version bump only for package @elux/taro
+
+
+
+
+
+# [2.6.0-alpha.1](https://github.com/hiisea/elux/compare/v2.6.0-alpha.0...v2.6.0-alpha.1) (2022-08-01)
+
+**Note:** Version bump only for package @elux/taro
+
+
+
+
+
+# [2.6.0-alpha.0](https://github.com/hiisea/elux/compare/v2.5.1...v2.6.0-alpha.0) (2022-07-27)
+
+**Note:** Version bump only for package @elux/taro
+
+
+
+
+
 # [2.4.0](https://github.com/hiisea/elux/compare/v2.3.4...v2.4.0) (2022-07-16)
 
 **Note:** Version bump only for package @elux/taro