+++ packages/vue-web/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/hiisea/elux/compare/v2.6.0-alpha.4...v2.6.0) (2022-08-02)
+
+**Note:** Version bump only for package @elux/vue-web
+
+
+
+
+
+# [2.6.0-alpha.2](https://github.com/hiisea/elux/compare/v2.6.0-alpha.1...v2.6.0-alpha.2) (2022-08-02)
+
+**Note:** Version bump only for package @elux/vue-web
+
+
+
+
+
+# [2.6.0-alpha.1](https://github.com/hiisea/elux/compare/v2.6.0-alpha.0...v2.6.0-alpha.1) (2022-08-01)
+
+
+### Features
+
+* 拆分与UI框架无关的model专用包 ([b2a2809](https://github.com/hiisea/elux/commit/b2a2809ea10f29af24fb70d4a3a83d82564ddbef))
+
+
+
+
+
+# [2.6.0-alpha.0](https://github.com/hiisea/elux/compare/v2.5.1...v2.6.0-alpha.0) (2022-07-27)
+
+
+### Features
+
+* 支持统一UI、纯Model ([3c6c083](https://github.com/hiisea/elux/commit/3c6c0832626e92277d214b4e6633e79d9b13d03c))
+
+
+
+
+
 ## [2.5.1](https://github.com/hiisea/elux/compare/v2.5.0...v2.5.1) (2022-07-20)